# perf: optimize Document init