--- a/docarray/document/mixins/property.py
+++ b/docarray/document/mixins/property.py
@@ -4,7 +4,8 @@ from typing import TYPE_CHECKING, Optional
 from ._property import _PropertyMixin
 
 if TYPE_CHECKING:
-    from ...types import DocumentContentType
+    from ...types import DocumentContentType, ArrayType
+    from ... import DocumentArray
 
 _all_mime_types = set(mimetypes.types_map.values())
 
@@ -22,16 +23,66 @@ class PropertyMixin(_PropertyMixin):
         if ct:
             return getattr(self, ct)
 
+    @_PropertyMixin.text.setter
+    def text(self, value: str):
+        self._clear_content()
+        self._data.text = value
+
+    @_PropertyMixin.blob.setter
+    def blob(self, value: bytes):
+        self._clear_content()
+        self._data.blob = value
+
+    @_PropertyMixin.tensor.setter
+    def tensor(self, value: 'ArrayType'):
+        self._clear_content()
+        self._data.tensor = value
+
     @content.setter
     def content(self, value: 'DocumentContentType'):
         if value is None:
             self._clear_content()
         elif isinstance(value, bytes):
-            self.blob = value
+            self._data.blob = value
         elif isinstance(value, str):
-            self.text = value
+            self._data.text = value
         else:
-            self.tensor = value
+            self._data.tensor = value
+
+    @_PropertyMixin.uri.setter
+    def uri(self, value: str):
+        mime_type = mimetypes.guess_type(value)[0]
+
+        if mime_type:
+            self._data.mime_type = mime_type
+        self._data.uri = value
+
+    @_PropertyMixin.mime_type.setter
+    def mime_type(self, value: str):
+        if value not in _all_mime_types:
+            # given but not recognizable, do best guess
+            r = mimetypes.guess_type(f'*.{value}')[0]
+            value = r or value
+
+        self._data.mime_type = value
+
+    @_PropertyMixin.chunks.setter
+    def chunks(self, value: 'DocumentArray'):
+        from ...array.chunk import ChunkArray
+
+        if not isinstance(value, ChunkArray):
+            value = ChunkArray(value, reference_doc=self._data._reference_doc)
+
+        self._data.chunks = value
+
+    @_PropertyMixin.matches.setter
+    def matches(self, value: 'DocumentArray'):
+        from ...array.match import MatchArray
+
+        if not isinstance(value, MatchArray):
+            value = MatchArray(value, reference_doc=self._data._reference_doc)
+
+        self._data.matches = value
 
     @property
     def content_type(self) -> Optional[str]: