--- a/docarray/score/__init__.py
+++ b/docarray/score/__init__.py
@@ -5,3 +5,4 @@ from ..base import BaseDCType
 
 class NamedScore(AllMixins, BaseDCType):
     _data_class = NamedScoreData
+    _post_init_fields = ()