Init 300K classes, Py37, MacM1

Before:
```
init empty Python class ...	0.028s
init Doc no content ...	2.251s
init Doc with content ...	3.091s
init Doc with content and embedding ...	3.463s
```

This PR
```
init empty Python class ...	0.029s
init Doc no content ...	1.025s
init Doc with content ...	1.268s
init Doc with content and embedding ...	1.361s
```

---

```python
from docarray import Document
from toytime import TimeContext


class A:
    ...


with TimeContext('init empty Python class'):
    for _ in range(300_000):
        A()

with TimeContext('init no Doc content'):
    for _ in range(300_000):
        Document()

with TimeContext('init with Doc content'):
    for _ in range(300_000):
        Document(content='hello')

with TimeContext('init with Doc content and embedding'):
    for _ in range(300_000):
        Document(content='hello', embedding=[1, 2, 3], parent_id='213')

```